+++ tests/src/rules/no-duplicates.js
@@ -169,6 +169,12 @@ ruleTester.run('no-duplicates', rule, {
     }),
 
     test({
+      code: "import def from './foo'; import {x} from './foo'",
+      output: "import def, {x} from './foo'; ",
+      errors: ['\'./foo\' imported multiple times.', '\'./foo\' imported multiple times.'],
+    }),
+
+    test({
       code: "import {x} from './foo'; import def from './foo'",
       output: "import def, {x} from './foo'; ",
       errors: ['\'./foo\' imported multiple times.', '\'./foo\' imported multiple times.'],